Refactor MI credential tests into dedicated SQL-Server-free unit test class
Add R_ManagedIdentityTokenAcquisitionError resource key for the token
acquisition exception path and move all Mockito-based pure unit tests from
SQLServerConnectionTest into a new SQLServerConnectionUnitTest class that does
not require a live SQL Server connection.
